requirements.txt
ccxt==4.4.42
pandas==2.3.0
numpy==2.0.2
python-telegram-bot==21.9
openpyxl==3.1.5
python-dotenv~=1.1.0